requirements.txt
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-WebEngine==6.7.0
pyperclip==1.9.0
pynput==1.7.7
cryptography==41.0.7
python-dotenv==1.0.0
matplotlib==3.8.0
jsonschema==4.17.0
